+++ ostap/math/tests/test_math_carlson.py
@@ -485,14 +485,30 @@ def test_carlson_KmE ( ) :
 #  @see https://indico.cern.ch/event/368497/contributions/1786992/attachments/1134067/1621999/davydychev.PDF
 #  @see http://cds.cern.ch/record/583358/files/0209233.pdf
 #  @see https://www.researchgate.net/publication/2054534_Three-body_phase_space_symmetrical_treatments
+#
+#  @see A.Davydychev and R.Delbourgo,
+#       "Explicitly symmetrical treatment of three body phase space",
+#       J.Phys. A37 (2004) 4871, arXiv:hep-th/0311075",
+#       doi = 10.1088/0305-4470/37/17/016
+#  @see https://arxiv.org/abs/hep-th/0311075
+#  @see https://iopscience.iop.org/article/10.1088/0305-4470/37/17/016
+#
 def test_carlson_PS3 ( ) :
     """Test 3-body phase space calculation via elliptic integrals
+    
     - see Ostap.Math.PhaseSpace3
     - see Ostap.Math.PhaseSpace3s
+    
     - see Ostap.Kinematics.phasespace3
     - see https://indico.cern.ch/event/368497/contributions/1786992/attachments/1134067/1621999/davydychev.PDF
     - see http://cds.cern.ch/record/583358/files/0209233.pdf
     - see https://www.researchgate.net/publication/2054534_Three-body_phase_space_symmetrical_treatments
+    
+    - see A.Davydychev and R.Delbourgo, ``Explicitly symmetrical treatment of three body phase space'',
+    J.Phys. A37 (2004) 4871, arXiv:hep-th/0311075,
+    doi = 10.1088/0305-4470/37/17/016
+    - see https://arxiv.org/abs/hep-th/0311075
+    - see https://iopscience.iop.org/article/10.1088/0305-4470/37/17/016    
     """
     logger = getLogger( 'test_carlson_PS3')
     logger.info ( 'Test 3-body phase space calculation via elliptic integrals' )